Support 64-bit (very-large) allocation sizes.

+++ blitz.mod/blitz_memory.c
@@ -9,7 +9,7 @@
 //Used by (some) GC routines to allocate large chunks, and GC mem needs to be on 16 byte boundary for
 //mem bit flags system in ref couter/Mark Sibly GCs...
 //
-void *bbMemAlloc( int size ){
+void *bbMemAlloc( size_t size ){
 	void *p;
 	
 	//p=malloc( size );
@@ -22,7 +22,7 @@ void bbMemFree( void *p ){
 	if( p ) GC_free( p );
 }
 
-void *bbMemExtend( void *mem,int size,int new_size ){
+void *bbMemExtend( void *mem,size_t size,size_t new_size ){
 	void *p;
 	p=bbMemAlloc( new_size );
 	bbMemCopy( p,mem,size );
@@ -30,14 +30,14 @@ void *bbMemExtend( void *mem,int size,int new_size ){
 	return p;
 }
 
-void bbMemClear( void *dst,int size ){
+void bbMemClear( void *dst,size_t size ){
 	memset( dst,0,size );
 }
 
-void bbMemCopy( void *dst,const void *src,int size ){
+void bbMemCopy( void *dst,const void *src,size_t size ){
 	memcpy( dst,src,size );
 }
 
-void bbMemMove( void *dst,const void *src,int size ){
+void bbMemMove( void *dst,const void *src,size_t size ){
 	memmove( dst,src,size );
 }
